+++ src/android/com/eclipsesource/firebase/analytics/AnalyticsHandler.kt
@@ -8,20 +8,22 @@ import com.eclipsesource.v8.V8Object
 @Suppress("PARAMETER_NAME_CHANGED_ON_OVERRIDE")
 class AnalyticsHandler(private val scope: ActivityScope) : ObjectHandler<Analytics> {
 
-  override val type = "com.eclipsesource.firebase.Analytics"
+    override val type = "com.eclipsesource.firebase.Analytics"
 
-  override val properties = listOf<Property<Analytics, *>>(
-      BooleanProperty("analyticsCollectionEnabled", { setAnalyticsCollectionEnabled(it ?: false) }),
-      StringProperty("screenName", { setScreenName(it) }),
-      StringProperty("userId", { setUserId(it.orEmpty()) })
-  )
+    override val properties = listOf<Property<Analytics, *>>(
+            BooleanProperty("analyticsCollectionEnabled", {
+                setAnalyticsCollectionEnabled(it ?: false)
+            }),
+            StringProperty("screenName", { setScreenName(it) }),
+            StringProperty("userId", { setUserId(it.orEmpty()) })
+    )
 
-  override fun create(id: String, properties: V8Object) = Analytics(scope.activity)
+    override fun create(id: String, properties: V8Object) = Analytics(scope.activity)
 
-  override fun call(analytics: Analytics, method: String, properties: V8Object) = when (method) {
-    "logEvent" -> analytics.logEvent(properties.getString("name"), properties.getObjectOrNull("data"))
-    "setUserProperty" -> analytics.setUserProperty(properties.getString("key"), properties.getStringOrNull("value"))
-    else -> null
-  }
+    override fun call(analytics: Analytics, method: String, properties: V8Object) = when (method) {
+        "logEvent" -> analytics.logEvent(properties.getString("name"), properties.getObjectOrNull("data"))
+        "setUserProperty" -> analytics.setUserProperty(properties.getString("key"), properties.getStringOrNull("value"))
+        else -> null
+    }
 
 }