+++ types/index.d.ts
@@ -1,7 +1,37 @@
-import { EventObject, NativeObject, PropertyChangedEvent, Widget, WidgetEvents, WidgetProperties } from 'tabris';
+import { EventObject, NativeObject, PropertyChangedEvent, Widget } from 'tabris';
 
 declare global {
   namespace firebase {
+    const Performance: Performance;
+    interface Performance extends NativeObject {
+      performanceCollectionEnabled: boolean;
+      startTrace(name: string): void;
+      stopTrace(name: string): void;
+      incrementMetrics(name1: string, name2: string): void;
+      incrementMetrics(name1: string, name2: string, value: number): void;
+    }
+    const Crashlytics: Crashlytics;
+    interface Crashlytics extends NativeObject {
+      setCrashlyticsCollectionEnabled(enabled: boolean): void;
+      setUserIdentifier(id: string): void;
+      makeCrash(): void;
+      log(value: string): void;
+      log(priority: 0 | 1 | 2 | 3 | 4, tag: string, value: string): void;
+      // Customizing
+      setBool(key: string, value: boolean): void;
+      setString(key: string, value: string): void;
+      setInt(key: string, value: number): void;
+    }
+    const RemoteConfig: RemoteConfig;
+    interface RemoteConfig extends NativeObject {
+      minimumFetchIntervalInSeconds: number;
+      setDefaults(values: Object): void;
+      getBoolean(key: string): boolean;
+      getDouble(key: string): number;
+      getLong(key: string): number;
+      getString(key: string): string;
+    }
+
     const Analytics: Analytics;
     const Messaging: Messaging;
     const MessagingEvents: MessagingEvents;
@@ -53,4 +83,4 @@ declare global {
   }
 }
 
-export {};
+export { };